Load marker details on demand
