+++ tests/test-recommendations.js
@@ -48,7 +48,7 @@ const _realFetch = globalThis.fetch;
   const categoryPageViewSrc = await fetchWithRetry('js/category-page-view.js');
   const categoryViewRenderersSrc = await fetchWithRetry('js/category-view-renderers.js');
   const chartCardRecsSrc = await fetchWithRetry('js/chart-card-recs.js');
-  const markerDetailSrc = await fetchWithRetry('js/marker-detail-modal.js');
+  const markerDetailSrc = await fetchWithRetry('js/marker-detail-modal-impl.js');
   const dashboardWidgetsSrc = await fetchWithRetry('js/dashboard-widgets.js');
   const dashboardWidgetRenderersSrc = await fetchWithRetry('js/dashboard-widget-renderers.js');
   const dashboardRecommendationWidgetSrc = await fetchWithRetry('js/dashboard-recommendation-widget.js');
