+++ tsconfig.checkjs.json
@@ -190,7 +190,9 @@
     "js/light-tools.js",
     "js/main.js",
     "js/marker-detail-modal.js",
+    "js/marker-detail-modal-impl.js",
     "js/marker-detail-runtime.js",
+    "js/modal-trigger-memory.js",
     "js/mobile-dashboard.js",
     "js/mobile-dashboard-runtime.js",
     "js/modal-lifecycle.js",
